requirements.txt
fastapi
uvicorn[standard]
requests
httpx
pydantic
torch
transformers
pillow
pytesseract
sentence-transformers